requirements.txt
bleach==6.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
kaggle==1.6.17
numpy==2.0.1
pandas==2.2.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
py4j==0.10.9.7
pyspark==3.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.32
text-unidecode==1.3
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
webencodings==0.5.1
